requirements.txt
weaviate-client==3.23.1
python-dotenv==1.0.0
openai==0.27.9
black==23.7.0
wasabi==1.1.2
fastapi==0.102.0
uvicorn[standard]
click== 8.1.7
pytest
setuptools
wheel
twine
spacy==3.6.1
tiktoken==0.5.1
sentence-transformers
transformers
torch
huggingface_hub
asyncio
accelerate
cohere==4.33
requests
pypdf2
pre-commit
ipython
httpx
atlassian
markdownify
